DJANGO/django_teamtest_제출/requirements.txt
django~=5.2.0
django-environ
djangorestframework~=3.16.0
django-cors-headers

pyhub-llm[all]

django-debug-toolbar

django-bootstrap5
django-extensions

pandas
openpyxl

pytest